Added PublicAppointmentSlotPicker file and Made the requested Changes
